% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{anova.gblup}
\alias{anova.gblup}
\title{print \code{\link{gblup}} object}
\usage{
\method{anova}{gblup}(gblup)
}
\arguments{
\item{gblup}{an object of the class gblup or summary.gblup generated by \code{\link{gblup}}}
}
\value{
a dataset with one row per fixed effect and columns for estimated value, S.E., test statistic and p-value
}
\description{
A prints a table of estimated fixed effects and tests for gblup or summary.gblup objects
}
\seealso{
\code{\link{gblup}}
}

