% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\docType{data}
\name{araGenes}
\alias{araGenes}
\title{A nicely formatted GRanges object}
\format{An object of class \code{GRanges} of length 32556.}
\usage{
araGenes
}
\description{
A nicely formatted GRanges object
}
\keyword{datasets}
