% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwar.R
\docType{data}
\name{araGenome}
\alias{araGenome}
\title{The araGenome object.}
\format{An object of class \code{data.frame} with 32841 rows and 7 columns.}
\usage{
araGenome
}
\description{
The araGenome object.
}
\keyword{datasets}
