% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/additional.R
\name{diagprod}
\alias{diagprod}
\title{Efficient computation of diagonals of matrix products}
\usage{
diagprod(x, y)
}
\arguments{
\item{x}{A matrix of size n by m}

\item{y}{A symetric matrix of size m by m}
}
\value{
A vector of length n
}
\description{
efficiently calculates the diagonal of  \code{x*y*t(x)}.
}

