% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwar.R
\name{expression_by_snp}
\alias{expression_by_snp}
\title{Based on a GWAS table and a rank, returns a table of expression values for a gene where accessions that contain the SNP have TRUE in SNP.
If no GeneID is specified, the gene that is closest to the SNP will be looked up.}
\usage{
expression_by_snp(
  gwas_table,
  SNPrank,
  GeneID = NULL,
  plot = FALSE,
  nobees = FALSE,
  SNPmatrix = NULL,
  ...
)
}
\arguments{
\item{gwas_table}{Object returned from \code{\link{format_gwas}} function}

\item{SNPrank}{The (-log10(p)) rank of the SNP of interest}

\item{GeneID}{If NULL (default), retrieves counts for the gene that is nearest. If not NULL it should be a TAIR GeneID ("ATXGNNNNN"). If a GeneID is provided expression of that gene will be intersected with the SNP presence.}

\item{plot}{If TRUE, will return a plot instead of a table.}

\item{nobees}{If TRUE, observations will be plotted with geom_point.}

\item{SNPmatrix}{SNPmatrix, to control if accessions are based on 1001genomes.org or local matrix.}
}
\description{
Based on a GWAS table and a rank, returns a table of expression values for a gene where accessions that contain the SNP have TRUE in SNP.
If no GeneID is specified, the gene that is closest to the SNP will be looked up.
}
\seealso{
\code{\link{format_gwas}}

\code{\link{get_expression}}

\code{\link{get_nearest_genes}}
}
