% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{find_peaks}
\alias{find_peaks}
\title{a function to confirm GWA peaks and to detect secondary peaks within chromosomes}
\usage{
find_peaks(gbl, gwa, Z, map, threshold = 0.95, criteria = c("qvalue",
  "bonferroni", "uncorrected"))
}
\arguments{
\item{gbl}{a gblup object generated by \code{\link{gblup}}}

\item{gwa}{a gblup object generated by \code{\link{gwa}}}

\item{Z}{a matrix of standardized genotypes}

\item{map}{a map dataset for the SNP in Z}

\item{threshold}{a significance threshold for GWA}

\item{criteria}{the p-value correction to use}
}
\value{
list with three elements 
\itemize{ 
  \item {\code{gblup}} {the new gblup object with peak SNP as fixed effects} 
  \item{\code{gwas}} {a new gwas objectest to look for peaks}
  \item{\code{peaks}} {a map with the positions of the included peaks}
}
}
\description{
A function to find secondary peaks in a GWA
}

