% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{format_gwas}
\alias{format_gwas}
\title{This function has no real purpose.}
\usage{
format_gwas(gwas)
}
\arguments{
\item{gwas}{A gwas result table}
}
\description{
This function has no real purpose.
}
