% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{format_gwasflow}
\alias{format_gwasflow}
\title{Read the GWAS tables produced by gwas-flow}
\usage{
format_gwasflow(gwasflow_out, permutation_results = NULL)
}
\arguments{
\item{gwasflow_out}{Results table (csv)}

\item{permutation_results}{Results from permutation test (if done)}
}
\description{
Read the GWAS tables produced by gwas-flow
}
\details{
A special case of the \code{\link{read_gwas}} function
}
\seealso{
\code{\link{read_gwas}}
}
