% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{format_sommer_gwas}
\alias{format_sommer_gwas}
\title{Read a GWAS object produced by the sommer package.}
\usage{
format_sommer_gwas(sommer_output)
}
\arguments{
\item{sommer_output}{output of the sommer::gwas() function.}
}
\description{
Read a GWAS object produced by the sommer package.
}
\details{
This function is a flexible version of \code{\link{read_gwas}}
}
