% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gblup.R
\name{gblup.default}
\alias{gblup.default}
\title{@rdname gblup
 @export}
\usage{
\method{gblup}{default}(rsp, data, design, G, vdata = NULL, wt = NULL, ...)
}
\description{
@rdname gblup
 @export
}

