% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gblup.R
\name{gblup}
\alias{gblup}
\alias{gblup.default}
\title{prediction of genomic breeding values for use in GWA}
\usage{
gblup(rsp, data, design, G, vdata = NULL, wt = NULL, ...)

\method{gblup}{default}(rsp, data, design, G, vdata = NULL, wt = NULL, ...)
}
\arguments{
\item{rsp}{a string with the name of the response variable if a vector, first element is used.}

\item{data}{either a gpData object from synbreed or a dataframe containing records for response variable and classification factors used as fixed or random effects}

\item{design}{a list or a vector with formulas for fixed and random effects}

\item{G}{a proportional covariance for genetics effects being predicted.}

\item{vdata}{a list with proportional covariance matrices of additional random effects (not included in data)}

\item{wt}{a vector of weigths (proportional to diagonals of residual covariance matrix), or a matrix of weights with anmimals in rows and response variables in columns}

\item{...}{additional parameters for regress function}

\item{vdata}{a list with proportional covariance matrices of additional random effects (not included in data)}
}
\value{
an object of the class gblup: a list the following components
\itemize{ 
   \item {\code{name}} {trait name} 
   \item{\code{llik}} {log-likelihood value at parameter estimates}
   \item{\code{cycle}} {number of iterations to convergence}
   \item{\code{Q}} {a matrix: see regress}
   \item{\code{V}} {phenotypic covariance matrix}
   \item{\code{Vinv}} {inverse of V}
   \item{\code{sigma}} {vector of estimated variance components}
   \item{\code{coefm}} {matrix of estimated fixed effects coefficients }
   \item{\code{model}} {the fit model}
   \item{\code{ehat}} {vector of residuals}
   \item{\code{pos}} {see regress}
}
}
\description{
A class for GBLUP prediction to implement GWA.
}

