% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwar.R
\name{get_accessions}
\alias{get_accessions}
\title{Obtain accession ids that carry a SNP.
This function takes a table of gwas pvalues per chromosome and numeric position and retrieves the accessions that carry that SNP.}
\usage{
get_accessions(gwas_table, SNPrank, SNPmatrix = NULL)
}
\arguments{
\item{gwas_table}{Object returned from \code{\link{format_gwas}} function}

\item{SNPrank}{The (-log10(p)) rank of the SNP of interest}

\item{SNPmatrix}{A path pointing to a SNPmatrix in .fst format.}
}
\description{
Obtain accession ids that carry a SNP.
This function takes a table of gwas pvalues per chromosome and numeric position and retrieves the accessions that carry that SNP.
}
\details{
If no SNPmatrix is provided, the information will be obtained from 1001genomes.org

This function returns a single column named ACC_ID, that contains the accessions that contain the snp.
}
