% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{get_expression}
\alias{get_expression}
\title{Query the AraPheno API to get expression data for a gene of interest}
\usage{
get_expression(GeneID)
}
\arguments{
\item{GeneID}{An Arabidopsis gene identifier}
}
\description{
Query the AraPheno API to get expression data for a gene of interest
}
\examples{
get_expression("AT4G21940")
}
