% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{get_nearest_genes}
\alias{get_nearest_genes}
\title{Find genes that are closest to a SNP.}
\usage{
get_nearest_genes(gwas_table = NULL, n_hit = 1)
}
\arguments{
\item{gwas_table}{Object returned from \code\link{read_gwas()} function}

\item{n_hit}{The number of SNPs that should be looked up.}
}
\description{
Find genes that are closest to a SNP.
}
\details{
Based on a GWAS table, returns the gene annotation that is closest to each SNP for the number of SNP specified. This function always returns the closest annotation. To limit the lookup range use {/link get_overlapping_genes()}  Lookup is done via ensembl plants; requires internet connection.
}
\seealso{
\code{\link{read_gwas}}

\code{\link{get_overlapping_genes}}
}
