% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{get_overlapping_genes}
\alias{get_overlapping_genes}
\title{Based on a GWAS table, returns the gene annotation that is overlapping with each SNP for the number of SNP specified.
If there is no overlap, the SNP is returned, but the GeneID column is empty.
Lookup is done via ensembl plants; requires internet connection.
!This function will assign araGenes in the Global Environment!}
\usage{
get_overlapping_genes(gwas_table = NULL, n_hit = 1, distance = -1)
}
\arguments{
\item{gwas_table}{Object returned from read_gwas() function}

\item{n_hit}{The number of SNPs that should be looked up.}

\item{distance}{The maximum distance from SNP to annotation, can be varied to lookup genes within a specific distance}
}
\description{
Based on a GWAS table, returns the gene annotation that is overlapping with each SNP for the number of SNP specified.
If there is no overlap, the SNP is returned, but the GeneID column is empty.
Lookup is done via ensembl plants; requires internet connection.
!This function will assign araGenes in the Global Environment!
}
\seealso{
\code{\link{read_gwas}}

\code{\link{get_nearest_genes}}
}
