% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{get_polymorph_acc}
\alias{get_polymorph_acc}
\title{Query the 1001genomes API to obtain accession ids that carry a SNP of interest}
\usage{
get_polymorph_acc(gwas_table, SNPrank)
}
\arguments{
\item{gwas_table}{Object returned from read_gwas() function}

\item{SNPrank}{The (-log10(p)) rank of the SNP of interest}
}
\description{
Query the 1001genomes API to obtain accession ids that carry a SNP of interest
}
\seealso{
\code{\link{read_gwas}}
}
