% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{get_SNPmatrix_acc}
\alias{get_SNPmatrix_acc}
\title{Obtain accession ids that carry a SNP of interest from a local SNPmatrix. This is a sister of get_polymorph_acc}
\usage{
get_SNPmatrix_acc(gwas_table, SNPrank, SNPmatrix)
}
\arguments{
\item{gwas_table}{Object returned from \code{\link{read_gwas}} function}

\item{SNPrank}{The (-log10(p)) rank of the SNP of interest}

\item{SNPmatrix}{The SNPmatrix to use for identifying accessions that carry the relevant SNP.}
}
\description{
Obtain accession ids that carry a SNP of interest from a local SNPmatrix. This is a sister of get_polymorph_acc
}
\seealso{
\code{\link{read_gwas}}

\code{\link{get_polymorph_acc}}
}
