\name{gwaR-package}
\alias{gwaR-package}
\alias{gwaR}
\docType{package}
\title{
\packageTitle{gwaR}
}
\description{
\packageDescription{gwaR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gwaR}
\packageIndices{gwaR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gwaR}

Maintainer: \packageMaintainer{gwaR}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
