% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS.R
\name{gwas}
\alias{gwas}
\title{Estimation of SNP effects and their variances}
\usage{
gwas(gblup, x)
}
\arguments{
\item{gblup}{An object of class gblup}

\item{x}{A matrix of standardized genotypes with animals in columns and markers in rows}
}
\value{
an object of the class GWAS: a list with two vectors
\itemize{ 
   \item {\code{ghat}} {estimated SNP effects} 
   \item{\code{varg}} {estimated SNP effect variances}
}
 @seealso \code{\link{summary.gwas}} \code{\link{plot.gwas}}
 @export
}
\description{
A class for GWA from a GBLUP object.
}

