% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{IC_base}
\alias{IC_base}
\title{Not intended for direct use. See  \code{\link{IC_gwa}} for details}
\usage{
IC_base(gbl, Z = NULL, map = NULL)
}
\arguments{
\item{gbl}{a gblup object generated by \code{\link{gblup}}}

\item{Z}{a matrix of standardized genotypes}

\item{map}{a map dataset for the SNP in Z}
}
\value{
the position of two peaks
}
\description{
A function used internally to compute Confidence Intervals of GWA peaks
}

