% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{IC_gwa}
\alias{IC_gwa}
\title{Computes confidence interval of GWA peak}
\usage{
IC_gwa(gbl, Z = NULL, map = NULL, rng = NULL, peak = 0, alpha = 0.99,
  iter = 500, NP = F)
}
\arguments{
\item{gbl}{a gblup object generated by \code{\link{gblup}}}

\item{Z}{a matrix of standardized genotypes}

\item{map}{a map dataset for the SNP in Z}

\item{rng}{a range of positions for which to compute the CI: a vector with three elements: chr, start, end}

\item{peak}{the position of the GWA peak should be within the range}

\item{alpha}{the confidence level for the confidence interval}

\item{iter}{the number of iterations}

\item{NP}{a logical value to compute non-parametric estimate}
}
\value{
a data.frame with 3 rows: first row is corresponds to the lower limit, the central row corresponds to the peak and the 3rd row is the lower limit. the columns correspond to: a)  name of closest snp, b) SNP chromosome, c) SNP position, D) exact position
}
\description{
A function usedto compute Confidence Intervals of GWA peaks
}

