% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{intersect_expression_snp}
\alias{intersect_expression_snp}
\title{Based on a GeneID and a GWAS table and a rank, returns a table of expression values for that gene,
where accessions that contain the SNP have TRUE in hasSNP}
\usage{
intersect_expression_snp(GeneID, gwas_table, SNPrank)
}
\arguments{
\item{GeneID}{An Arabidopsis thaliana gene identifier}

\item{gwas_table}{Object returned from \code\link{read_gwas()} function}

\item{SNPrank}{The (-log10(p)) rank of the SNP of interest}
}
\description{
Based on a GeneID and a GWAS table and a rank, returns a table of expression values for that gene,
where accessions that contain the SNP have TRUE in hasSNP
}
\seealso{
\code{\link{get_expression}}

\code{\link{read_gwas}}
}
