% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{intersect_phenotype_snp}
\alias{intersect_phenotype_snp}
\title{Get intersection between phenotype, and variant table.}
\usage{
intersect_phenotype_snp(phenotype_table, phenotype, gwas_table, SNPrank,
  acc_col = "ACC_ID", specific = NULL)
}
\arguments{
\item{phenotype_table}{a table containing phenotyping measurements, and accession ids (see below)}

\item{phenotype}{a specific phenotype from the phenotype table. Must match to a column name of the phenotype table}

\item{gwas_table}{Object returned from \code\link{read_gwas()} function}

\item{SNPrank}{The (-log10(p)) rank of the SNP of interest}

\item{acc_col}{the column that contains accession identifiers.}

\item{specific}{(optional) treatment column that was used to split samples for specific GWAS.}
}
\description{
Get intersection between phenotype, and variant table.
}
\details{
Based on a table of phenotypes, a phenotype name, a GWAS table and a rank, returns a table of phenotype values for that gene, where accessions that contain the SNP have TRUE in hasSNP
}
\seealso{
\code{\link{get_phenotype}}

\code{\link{read_gwas}}

\code{\link{intersect_expression_snp}}
}
