% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwar.R
\name{linkage_phenotypes}
\alias{linkage_phenotypes}
\title{Calculate linkage around SNP of interest and obtain phenotypes from phenotype table.}
\usage{
linkage_phenotypes(
  phenotype_table,
  phenotype,
  chrom,
  pos,
  nuc_range,
  ld_stats = "LLR",
  ld_cutoff = 10,
  SNPmatrix = "~/SNPmatrix.fst",
  specific = c("bx", "dag"),
  impacts = c("MODERATE", "HIGH")
)
}
\arguments{
\item{phenotype_table}{A table containing phenotypes that should be retrieved. Column with accession IDs has to be name 'ACC_ID'}

\item{phenotype}{Name of the phenotype of interest}

\item{chrom}{Chromosome where the SNP of interest is}

\item{pos}{Position of the SNP of interest on the chromosome}

\item{nuc_range}{Range of nucleotides that will be analyzed (total, split evenly up and downstream of the SNP)}

\item{ld_stats}{The LD statistics, see SNPStats::ld. Default is LLR}

\item{ld_cutoff}{Only SNPs that have an LD values >= this will be plotted (default 10)}

\item{SNPmatrix}{default "~/SNPmatrix.fst". If Needs to provided as a path to an fst file. Will be used for linkage calculation instead of polymorph. To use polymorph set to NULL}

\item{specific}{(optional) treatment column that was used to split samples for specific GWAS.}

\item{impacts}{SNP impacts to subset for can ("MODIFIER","LOW","MODERATE","HIGH"; default: c("MODERATE","HIGH"))}

\item{gwas_table}{Object returned from \code{\link{format_gwas}} function}
}
\description{
Calculate linkage around SNP of interest and obtain phenotypes from phenotype table.
}
\details{
This function calculates LD in a specific region.

The SNPs per accession in the region are retrieved from polymorph or from a provided SNPmatrix.

For all SNPs that pass some criteria (see params), the phenotypes are then retrieved and split by presence of each SNP (SNP = TRUE or FALSE).
}
