% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{lrt}
\alias{lrt}
\title{Likelihood ratio tests for GBLUP models}
\usage{
lrt(gb)
}
\arguments{
\item{gb}{a gblup model}
}
\value{
a list with elements
\itemize{ 
   \item {\code{pvalue}} {the pvalue of the LRT} 
   \item{\code{llik}} {Log-likelihood of full model, reduced model and their difference}
   \item{\code{VARS}} {a dataframe with variance component estimates under reduced and full model}
}
 @seealso \code{\link{gblup}} \code{\link{gwas}}
 @export
}
\description{
function for performing likelihood ratio test for  \code{h^2=0} of a \code{\link{gblup}} model
}

