% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/additional.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Manhattan plots of p-values}
\usage{
manhattan_plot(pvalues, map, threshold = 0.01, col = c("black", "red"), ...)
}
\arguments{
\item{pvalues}{A vector of pvalues with SNP, marker or position names}

\item{map}{a genetic map: a data.frame with SNP or marker names in rows, chromosome name (chr column) and position (pos column)}

\item{threshold}{the threshold of significance}

\item{col}{a vector of strings with color names for painting SNP according to their chromosome}

\item{...}{additional graphical parameters used by plot}
}
\value{
NULL. A graphic is be generated
}
\description{
A function to create manhattan plots from GWA and QTL mapping analyses.
}

