% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwar.R
\name{mine_gene}
\alias{mine_gene}
\title{Get information from thalemine.}
\usage{
mine_gene(GeneID)
}
\arguments{
\item{GeneID}{The GeneID of interest}
}
\description{
Get information from thalemine.
}
\details{
This retrieves information from the InterMineR API. Takes a geneID, returns publications, functional annotations and GO Term
}
