% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwar.R
\name{phenotype_by_snp}
\alias{phenotype_by_snp}
\title{Get intersection between phenotype, and SNP presence.}
\usage{
phenotype_by_snp(
  phenotype_table,
  phenotype,
  gwas_table,
  SNPrank,
  acc_col = "ACC_ID",
  specific = NULL,
  plot = FALSE,
  nobees = FALSE,
  SNPmatrix = NULL,
  ...
)
}
\arguments{
\item{phenotype_table}{a table containing phenotyping measurements, and accession ids (see below)}

\item{phenotype}{a specific phenotype from the phenotype table. Must match to a column name of the phenotype table}

\item{gwas_table}{Object returned from \code{\link{format_gwas}} function}

\item{SNPrank}{The (-log10(p)) rank of the SNP of interest}

\item{acc_col}{the column that contains accession identifiers.}

\item{specific}{(optional) treatment column that was used to split samples for specific GWAS.}

\item{plot}{Default: FALSE. Toggle if a plot should be returned}

\item{nobees}{Set to true to disable beeswarm geom (only relevant if plot = TRUE)}

\item{SNPmatrix, }{Provide SNPmatrix to control if accessions are based on 1001genomes.org or local matrix (default: NULL -> use 1001genomes).}

\item{...}{Arguments passed to other functions. Most relevant is}
}
\description{
Get intersection between phenotype, and SNP presence.
}
\details{
Based on a table of phenotypes, a phenotype name, a GWAS table and a rank, and a SNPmatrix returns a table of phenotype values for that gene, where accessions that contain the SNP have TRUE in SNP.

Also provides a simple default plot, toggled via the plot argument
}
\seealso{
\code{\link{get_phenotype}}

\code{\link{format_gwas}}

\code{\link{expression_by_snp}}

\code{\link{get_accessions}}
}
