% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/additional.R
\name{plot.gwas}
\alias{plot.gwas}
\title{Significance plots and pvalues from \code{\link{gwas}} output}
\usage{
\method{plot}{gwas}(gwas, correction = "bonf", gpdata = NULL,
  plotlog10 = FALSE, pvalue = 0.05, q.qplot = FALSE, ...)
}
\arguments{
\item{gwas}{an object of the class gwas generated by \code{\link{gwas}}}

\item{correction}{a string with value 'bonf' if compute bonferroni correction is needed}

\item{gpdata}{an object of class gpdata used to obtain a SNP map}

\item{plotlog10}{a logical value indicating if a manhattan plot is requested}

\item{pvalue}{significance threshold}

\item{q.qplot}{a logical value indicating if a q-qplot is requested}

\item{...}{}
}
\value{
a vector of pvalues with names equal to marker names. Optional output: Manhattan plot and q-qplot of pvalues
}
\description{
Computes pvalues from a GWA class object represents them graphically. using \code{\link{manhattan_plot}} and \code{\link{qqgplot}}
}

