% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{plot_anchored_ld_snpmatrix}
\alias{plot_anchored_ld_snpmatrix}
\title{Produce a plot of linked SNPs based on a SNPMatrix}
\usage{
plot_anchored_ld_snpmatrix(
  gwas_table,
  rank,
  nuc_range = 50000,
  snpmatrix_path = NULL,
  ld_stats = c("D.prime"),
  use_phenotype_table = NULL,
  acc_col = "ACC_ID",
  linkage_cutoff = 0.8,
  LD_legend = FALSE,
  data_only = FALSE
)
}
\arguments{
\item{gwas_table}{Object returned from \code{\link{read_gwas}} function}

\item{rank}{Rank of the SNP of interest}

\item{nuc_range}{Range of nucleotides that will be analyzed (total, split evenly up and downstream of the SNP)}

\item{snpmatrix_path}{filepath to SNPmatrix in fst format.}

\item{ld_stats}{The LD statistics, see SNPStats::ld}

\item{use_phenotype_table}{If supplied: Genotypes listed here will be used for linkage analysis. Otherwise, all accessions that carry this SNP will be included.}

\item{acc_col}{Name of the column containing accession identifiers in the phenotype table}

\item{linkage_cutoff}{Only SNPs that have an LD values >= this will be plotted (default 0.8)}

\item{LD_legend}{Toggle color legend for LD. Off by default.}

\item{data_only}{If true does not plot, only returns the data used for plotting.}
}
\description{
Produce a plot of linked SNPs based on a SNPMatrix
}
\details{
This function takes a gwas table and a rank and produces a plot that illustrates, SNPs that are within nuc_range around the SNP, the degree of linkage, the impact of the other SNPs and a track of gene annotations to easily identify linked SNPs that have a high impact.

Impact according to 1001genomes.org
}
