% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{plot_annotated_gwas}
\alias{plot_annotated_gwas}
\title{Plot annotated manhattan plot}
\usage{
plot_annotated_gwas(gwas_table, title = "No Title", subtitle = NULL,
  nlabels = 5, labeltype = "GeneId", match_nearest = FALSE,
  p_filter = 2, mac_filter = 0)
}
\arguments{
\item{gwas_table}{Object returned from read_gwas() function}

\item{title}{Specify plot title}

\item{subtitle}{Specify plot subtitle}

\item{nlabels}{How many labels should be plotted (default: 5)}

\item{labeltype}{Which entry from the lookup should be displayed (character, defaults to "GeneId"). Options: SNP_rank, chrom, pos, Significant, log10_p, mac, GeneId, max_distance, description, transcript_biotype, start_position, end_position.}

\item{match_nearest}{Defaults to FALSE; if TRUE will use find_nearest_genes instead of find_overlapping_genes for annotations}

\item{p_filter}{everything with a log10(p) below this value will not be included in the plot (default: 2)}

\item{mac_filter}{everything with a mac (minor allele count) below this will not be plotted.}
}
\description{
Plot annotated manhattan plot
}
\details{
Based on a GWAS table, generates a manhatten plot, with gene annotations. Number of annotations can be toggled by changing nlabels. By default only plots annotations for SNPs that are within a gene annotation. For performance reasons, everything with a log10(p) smaller than p_filter is filtered out.
}
\seealso{
\code{\link{read_gwas}}

\code{\link{plot_gwas}}

\code{\link{get_nearest_genes}}

\code{\link{get_overlapping_genes}}
}
