% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{plot_intersect_expression_snp}
\alias{plot_intersect_expression_snp}
\title{Based on a GWAS table and a rank, returns a table of expression values for the gene that is closest to that SNP
where accessions that contain the SNP have TRUE in hasSNP}
\usage{
plot_intersect_expression_snp(gwas_table, SNPrank, nobees = FALSE)
}
\arguments{
\item{gwas_table}{Object returned from \code\link{read_gwas()} function}

\item{SNPrank}{The (-log10(p)) rank of the SNP of interest}

\item{nobees}{Set to true to disable beeswarm geom}
}
\description{
Based on a GWAS table and a rank, returns a table of expression values for the gene that is closest to that SNP
where accessions that contain the SNP have TRUE in hasSNP
}
\seealso{
\code{\link{read_gwas}}

\code{\link{get_expression}}

\code{\link{get_nearest_genes}}

\code{\link{retrieve_counts}}
}
