% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{plot_intersect_phenotype_snp}
\alias{plot_intersect_phenotype_snp}
\title{Based on a GeneID and a GWAS table and a rank, produces a boxplot of that phenotype, grouped by presence of that SNP.}
\usage{
plot_intersect_phenotype_snp(phenotype_table, phenotype, gwas_table,
  SNPrank, acc_col = "ACC_ID")
}
\arguments{
\item{phenotype_table}{a table containing phenotyping measurements, and accession ids (see below)}

\item{phenotype}{a specific phenotype from the phenotype table. Must match to a column name of the phenotype table}

\item{gwas_table}{Object returned from read_gwas() function}

\item{SNPrank}{The (-log10(p)) rank of the SNP of interest}

\item{acc_col}{the column that contains accession identifiers.}
}
\description{
Based on a GeneID and a GWAS table and a rank, produces a boxplot of that phenotype, grouped by presence of that SNP.
}
\seealso{
\code{\link{get_phenotype}}

\code{\link{read_gwas}}

\code{\link{intersect_phenotype_snp}}
}
