% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{plot_intersect_phenotype_snpmatrix}
\alias{plot_intersect_phenotype_snpmatrix}
\title{Split phenotype table by SNP presence and plot}
\usage{
plot_intersect_phenotype_snpmatrix(
  phenotype_table,
  phenotype,
  gwas_table,
  SNPrank,
  SNPmatrix,
  acc_col = "ACC_ID",
  specific = NULL,
  nobees = FALSE
)
}
\arguments{
\item{phenotype_table}{a table containing phenotyping measurements, and accession ids (see below)}

\item{phenotype}{a specific phenotype from the phenotype table. Must match to a column name of the phenotype table}

\item{gwas_table}{Object returned from \code{\link{read_gwas}} function}

\item{SNPrank}{The (-log10(p)) rank of the SNP of interest}

\item{SNPmatrix}{The SNPmatrix to use for identifying accessions that carry the relevant SNP.}

\item{acc_col}{the column that contains accession identifiers.}

\item{specific}{(optional) treatment column that was used to split samples for specific GWAS.}

\item{nobees}{Set to true to disable beeswarm geom}
}
\description{
Split phenotype table by SNP presence and plot
}
\details{
Based on a Phenotype table, the name of the phenotype and a GWAS table and a rank, produces a boxplot of that phenotype, grouped by presence of that SNP.
}
\seealso{
\code{\link{get_phenotype}}

\code{\link{read_gwas}}

\code{\link{intersect_phenotype_snp}}
}
