% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.gblup}
\alias{print.gblup}
\title{print \code{\link{gblup}} object}
\usage{
\method{print}{gblup}(gblup)
}
\arguments{
\item{gblup}{an object of the class gblup generated by \code{\link{gblup}}}
}
\value{
Null. This function only prints to the screen:
}
\description{
A print method for gblup objects
}
\seealso{
\code{\link{gblup}}
}

