% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.summary.gblup}
\alias{print.summary.gblup}
\title{print \code{\link{summary.gblup}} object}
\usage{
\method{print}{summary.gblup}(gblup)
}
\arguments{
\item{gblup}{an object of the class symmary.gblup generated by \code{\link{summary.gblup}}}
}
\value{
Null. This function only prints to the screen:
}
\description{
A print method for summary.gblup objects
}
\seealso{
\code{\link{summary.gblup}}
}

