% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/additional.R
\name{qqgplot}
\alias{qqgplot}
\title{q-qplots of p-values}
\usage{
qqgplot(pvector, add = F, ...)
}
\arguments{
\item{pvector}{A vector of pvalues, assumed to follow a uniform distribution under the null hypothesis}

\item{add}{a logical value, T: adds to existing graphic or F: creates a new graphic}

\item{...}{additional parameters to parse to plot function}
}
\value{
NULL. A graphic will be generated
}
\description{
A function to create q-qplots of pvalues from high dimensional test (e.g: GWA, Differential expression analyses).
}

