% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwar.R
\name{read_limix}
\alias{read_limix}
\title{Read a GWAS result file from limix (csv) and put it into a specific (somewhat arbitrary) format.}
\usage{
read_limix(gwas_path)
}
\arguments{
\item{gwas_path}{Path to a gwas result file containing SNP positions and p-values}
}
\description{
Read a GWAS result file from limix (csv) and put it into a specific (somewhat arbitrary) format.
}
\details{
`format_gwas` returns a tibble containing SNP positions, -log10 transformed p-values and information whether a particular SNP passes a significance threshold after multiple testing correction (Bonferroni or Benjamini-Hochberg)
}
\seealso{
\code{\link{format_gwas}}
}
