% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{retrieve_counts}
\alias{retrieve_counts}
\title{Based on a GWAS table and a rank, returns a table of expression values for the gene that is closest to that SNP
where accessions that contain the SNP have TRUE in hasSNP}
\usage{
retrieve_counts(gwas_table, SNPrank)
}
\arguments{
\item{gwas_table}{Object returned from \code\link{read_gwas()} function}

\item{SNPrank}{The (-log10(p)) rank of the SNP of interest}
}
\description{
Based on a GWAS table and a rank, returns a table of expression values for the gene that is closest to that SNP
where accessions that contain the SNP have TRUE in hasSNP
}
\seealso{
\code{\link{read_gwas}}

\code{\link{get_expression}}

\code{\link{get_nearest_genes}}
}
