% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{run.gwa}
\alias{run.gwa}
\title{GBLUP and GWA computation}
\usage{
run.gwa(rsp, data, design, G, vdata = NULL, wt = NULL, x, LRT = F,
  threshold = 0.01, returnz = T, saveblup = F, basename = "", ...)
}
\arguments{
\item{rsp}{a string with the name of the response variable if a vector, first element is used.}

\item{data}{either a gpData object from synbreed or a dataframe containing records for response variable and classification factors used as fixed or random effects}

\item{design}{a list or a vector with formulas for fixed and random effects}

\item{G}{a proportional covariance for genetics effects being predicted.}

\item{vdata}{a list with proportional covariance matrices of additional random effects (not included in data)}

\item{wt}{a vector of weigths (proportional to diagonals of residual covariance matrix), or a matrix of weights with anmimals in rows and response variables in columns}

\item{x}{A matrix of standardized genotypes with animals in columns and markers in rows}

\item{LRT}{logical indicating if a LRT is performed to test for h2 before performing GWA}

\item{threshold}{the significance threshold of LRT}

\item{returnz}{a logical value. If T the t-statistics for each SNP is returned, if F estimated SNP effects and variances are returnes}

\item{...}{additional parameters for regress function}

\item{vdata}{a list with proportional covariance matrices of additional random effects (not included in data)}
}
\value{
if {\code{returnz=F}}, an object of the class GWAS: a list with two vectors
\itemize{ 
   \item {\code{ghat}} {estimated SNP effects} 
   \item{\code{varg}} {estimated SNP effect variances}
}

if {\code{returnz=F}} the t-statistic {\code{t=ghat/sqrt(varg)}} 
 @seealso \code{\link{gblup}} \code{\link{gwas}} \code{\link{lrt}}
 @export
}
\description{
function for fitting a  \code{\link{gblup}} model followed by \code{\link{gwas}} analysis.
}

