% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwar.R
\docType{data}
\name{sequenced_accessions}
\alias{sequenced_accessions}
\title{A list of sequenced accessions as of 2020}
\format{
A single column of accesion ids
\describe{
\item{ACC_ID}{Accession IDs}
}
}
\usage{
sequenced_accessions
}
\description{
A list of sequenced accessions as of 2020
}
\keyword{datasets}
