% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwar.R
\name{snp_linkage}
\alias{snp_linkage}
\title{Calculate linkage around SNP of interest.}
\usage{
snp_linkage(gwas_table, rank, anchored = FALSE, plot = FALSE,
  SNPmatrix = NULL, nuc_range = 50000, ld_depth = 1000,
  ld_stats = "R.squared", ld_symmetric = FALSE, ld_cutoff = 0.8,
  ld_legend = FALSE, use_phenotype_table = NULL, acc_col = "ACC_ID",
  use_all_acc = FALSE, data_only = FALSE)
}
\arguments{
\item{gwas_table}{Object returned from \code{\link{format_gwas}} function}

\item{rank}{Rank of the SNP of interest}

\item{anchored}{Perform anchored analysis. If TRUE Linkage will only be estimated for the SNP of interest, and the surrounding ones, but not between surrounding SNPs.}

\item{plot}{Default FALSE. Should a plot be returned, instead of the table? Only supported for anchored = T}

\item{SNPmatrix}{default NULL. Needs to provided as a path to an fst file. Will be used instead of polymorph.}

\item{nuc_range}{Range of nucleotides that will be analyzed (total, split evenly up and downstream of the SNP)}

\item{ld_depth}{Maximum SNP distance to calculate LD for (only relevant when anchored = FALSE)}

\item{ld_stats}{The LD statistics, see SNPStats::ld. Default is D.prime}

\item{ld_symmetric}{Should a symmetric matrix be returned? see SNPStats::ld}

\item{ld_cutoff}{Only SNPs that have an LD values >= this will be plotted (default 0.8)}

\item{ld_legend}{Toggle color legend for LD. Off by default.}

\item{use_phenotype_table}{If supplied: Genotypes listed here will be used for linkage analysis. Otherwise, all accessions that carry this SNP will be included.}

\item{acc_col}{Name of the column containing accession identifiers in the phenotype table}

\item{use_all_acc}{If true, will override use_phenotype_table and use all 1135 sequenced accessions.}

\item{data_only}{Return the data used for plotting, instead of the plot, defaults to FALSE.}
}
\description{
Calculate linkage around SNP of interest.
}
\details{
This function calculates LD in a specific region.

The SNPs per accession in the region are retrieved from polymorph or from a provided SNPmatrix.

Sometimes, the ld function produces no result, especially when used with anchored = T.

anchored analysis will use one SNP can compute LD of that SNP against a region.

Default settings return the LD table.
}
