% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{snp_linkage_snpmatrix}
\alias{snp_linkage_snpmatrix}
\title{Calculate linkage around SNP of interest.}
\usage{
snp_linkage_snpmatrix(
  gwas_table,
  rank,
  snpmatrix_path = NULL,
  nuc_range = 50000,
  ld_depth = 1000,
  ld_stats = "D.prime",
  ld_symmetric = FALSE,
  use_phenotype_table = NULL,
  acc_col = "ACC_ID",
  anchored = FALSE
)
}
\arguments{
\item{gwas_table}{Object returned from \code{\link{read_gwas}} function}

\item{rank}{Rank of the SNP of interest}

\item{snpmatrix_path}{the path to the SNPmatrix in fst format.}

\item{nuc_range}{Range of nucleotides that will be analyzed (total, split evenly up and downstream of the SNP)}

\item{ld_depth}{Maximum SNP distance to calculate LD for (only relevant when anchored = FALSE)}

\item{ld_stats}{The LD statistics, see SNPStats::ld. Default is D.prime}

\item{ld_symmetric}{Should a symmetric matrix be returned? see SNPStats::ld}

\item{use_phenotype_table}{If supplied: Genotypes listed here will be used for linkage analysis. Otherwise, all accessions in the SNPmatrix are used. Accessions are assumed to be named as numbers.}

\item{acc_col}{the column name of the accession column in the phenotype table}

\item{anchored}{Perform anchored analysis. If TRUE Linkage will only be estimated for the SNP of interest, and the surrounding ones, but not between surrounding SNPs.}
}
\description{
Calculate linkage around SNP of interest.
}
\details{
Details are fuzzy right now
}
