% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWAS.R
\name{summary.gblup}
\alias{summary.gblup}
\title{summary for \code{\link{gblup}} object}
\usage{
\method{summary}{gblup}(gblup, ehat = FALSE, sigma = FALSE)
}
\arguments{
\item{gblup}{an object of the class gblup generated by \code{\link{gblup}}}

\item{ehat}{a logical value to indicate if residuals should be computed}
}
\value{
a frequency table with the number of SNP significant at the thresholds specified by pvalue
\itemize{
   \item {\code{uhat}} {blup of individuals}
   \item{\code{sigma}} {estimated variance components and h2}
}
}
\description{
A summary function for gblup objects
}
\seealso{
\code{\link{gblup}}
}

