% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{summary.gwas}
\alias{summary.gwas}
\title{Significance summary for a \code{\link{gwas}} object}
\usage{
\method{summary}{gwas}(gwas, pvalue = c(0.001, 0.01, 0.025, 0.05, 0.1),
  correction = "bonf")
}
\arguments{
\item{gwas}{an object of the class gwas generated by \code{\link{gwas}}}

\item{pvalue}{a vector with the p-value thresholds to build frequency table}

\item{correction}{a string with value 'bonf' if compute bonferroni correction is needed}
}
\value{
a frequency table with the number of SNP significant at the thresholds specified by pvalue
}
\description{
A frequency table of number of significant tests at given thresholds
}

