% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{test.peak}
\alias{test.peak}
\title{Likelihood ratio tests for QTL peaks}
\usage{
test.peak(gb, x, peak_pos = NULL)
}
\arguments{
\item{gb}{a gblup model}

\item{x}{A matrix of standardized genotypes with animals in columns and markers in rows}

\item{peak_pos}{an index vector used to select the rows (markers) of x that capture the QTL peak. if NULL: the whole x matrix is used.}
}
\value{
a list with elements
\itemize{ 
   \item {\code{pvalue}} {the pvalue of the LRT} 
   \item{\code{llik}} {Log-likelihood of full model, reduced model and their difference}
   \item{\code{VARS}} {a dataframe with variance component estimates under reduced and full model}
}
}
\description{
function for performing likelihood ratio test  \code{h_peak^2=0} of QTL peak using a \code{\link{gblup}} model
}
\seealso{
\code{\link{gblup}} \code{\link{lrt}}
}

