% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwar.R
\docType{data}
\name{araGenes}
\alias{araGenes}
\title{araGenome as a nicely formatted GRanges object}
\format{An object of class \code{GRanges} of length 32556.}
\usage{
araGenes
}
\description{
araGenome as a nicely formatted GRanges object
}
\keyword{datasets}
