% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwar.R
\name{format_gwas}
\alias{format_gwas}
\title{Read an arbitrary GWAS table.}
\usage{
format_gwas(gwas_path, p_col = NULL, chrom_col = NULL,
  pos_col = NULL)
}
\arguments{
\item{gwas_path}{A gwas result table}

\item{p_col}{Name of the column containing p-values}

\item{chrom_col}{Name of the column containing chromosomes}

\item{pos_col}{Name of the column containing positions}
}
\description{
Read an arbitrary GWAS table.
}
\details{
This function is a flexible version of \code{\link{format_gwas}}
}
