% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwar.R
\name{get_expression}
\alias{get_expression}
\title{Get expression data for a gene of interest}
\usage{
get_expression(GeneID = NULL, study = 52, list_studies = FALSE)
}
\arguments{
\item{GeneID}{An Arabidopsis thaliana gene identifier}

\item{study}{(default 52) the study of interest at arapheno, see list here}

\item{list_studies}{If TRUE, will return a list of available studies from arapheno.}
}
\description{
Get expression data for a gene of interest
}
