% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{get_expression}
\alias{get_expression}
\title{Get expression data for a gene of interest}
\usage{
get_expression(GeneID)
}
\arguments{
\item{GeneID}{An Arabidopsis thaliana gene identifier}
}
\description{
Get expression data for a gene of interest
}
