% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/additional.R
\name{getpvalue}
\alias{getpvalue}
\title{pvalue computation for a \code{\link{gwas}} object}
\usage{
getpvalue(gwas, log.p = T)
}
\arguments{
\item{gwas}{an object of the class gwas generated by \code{\link{gwas}}}

\item{log.p}{a logical value to indicate if log10-pvalues are requested}
}
\value{
a vector of p-values or log10 p-values
}
\description{
Pvalue for blup objects
}

