% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/additional.R
\name{getpvalue}
\alias{getpvalue}
\title{Significance summary for a \code{\link{GWAS}} object}
\usage{
getpvalue(gwas, log.p = T)
}
\arguments{
\item{gwas}{an object of the class gwas generated by \code{\link{GWAS}}}
}
\value{
a vector of p-values
}
\description{
Pvalue for blup objects
}

