% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWAS.R
\name{gwas}
\alias{gwas}
\alias{gwas.default}
\title{Estimation of SNP effects and their variances}
\usage{
gwas(gblup, x)

\method{gwas}{default}(gblup, x)
}
\arguments{
\item{gblup}{An object of class gblup}

\item{x}{A matrix of standardized genotypes}
}
\value{
an object of the class GWAS: a list with two vectors
\itemize{
   \item {\code{ghat}} {estimated SNP effects}
   \item{\code{varg}} {estimated SNP effect variances}
}
}
\description{
A class for GWA from a GBLUP object.


}
\seealso{
\code{\link{summary.gwas}} \code{\link{plot.gwas}}
}

