% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{plot_acc_map}
\alias{plot_acc_map}
\title{Plot an interactive map of accessions that carry a SNP of interest}
\usage{
plot_acc_map(gwas_table, SNPrank)
}
\arguments{
\item{gwas_table}{Object returned from read_gwas() function}

\item{SNPrank}{The (-log10(p)) rank of the SNP of interest}
}
\description{
Plot an interactive map of accessions that carry a SNP of interest
}
\seealso{
[read_gwas()]
}
