% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{plot_gwas}
\alias{plot_gwas}
\title{Based on a GWAS table, generates a manhatten plot.}
\usage{
plot_gwas(
  gwas_table,
  title = "No Title",
  subtitle = NULL,
  p_filter = 2,
  mac_filter = 0
)
}
\arguments{
\item{gwas_table}{Object returned from \code{\link{read_gwas}} function}

\item{title}{Specify plot title}

\item{subtitle}{Specify plot subtitle}

\item{p_filter}{everything with a log10(p) below this value will not be included in the plot}

\item{mac_filter}{everything with a mac (minor allele count) below this will not be plotted. This requires a mac column.}
}
\description{
Based on a GWAS table, generates a manhatten plot.
}
\details{
For performance reasons, everything with a log10(p) smaller than p_filter is filtered out.
}
\seealso{
\code{\link{read_gwas}}

\code{\link{plot_annotated_gwas}}
}
