% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_functions.R
\name{snp_linkage}
\alias{snp_linkage}
\title{Calculate linkage around SNP of interest.}
\usage{
snp_linkage(gwas_table, rank, nuc_range = 50000, ld_depth = "1000",
  ld_stats = c("D.prime", "R.squared"), ld_symmetric = FALSE,
  use_phenotype_table = NULL, acc_col = "ACC_ID", anchored = FALSE,
  use_all_acc = FALSE)
}
\arguments{
\item{gwas_table}{Object returned from \code{\link{read_gwas}} function}

\item{rank}{Rank of the SNP of interest}

\item{nuc_range}{Range of nucleotides that will be analyzed (total, split evenly up and downstream of the SNP)}

\item{ld_depth}{Maximum SNP distance to calculate LD for (only relevant when anchored = FALSE)}

\item{ld_stats}{The LD statistics, see SNPStats::ld}

\item{ld_symmetric}{Should a symmetric matrix be returned? see SNPStats::ld}

\item{use_phenotype_table}{If supplied: Genotypes listed here will be used for linkage analysis. Otherwise, all accessions that carry this SNP will be included.}

\item{acc_col}{Name of the column containing accession identifiers in the phenotype table}

\item{anchored}{Perform anchored analysis. If TRUE Linkage will only be estimated for the SNP of interest, and the surrounding ones, but not between surrounding SNPs.}

\item{use_all_acc}{If true, will override use_phenotype_table and use all 1135 sequenced accessions.}
}
\description{
Calculate linkage around SNP of interest.
}
\details{
Details are fuzzy right now
}
