% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summary.gblup}
\alias{summary.gblup}
\title{summary for \code{\link{gblup}} object}
\usage{
\method{summary}{gblup}(gblup, ehat = FALSE, sigma = FALSE, fe = FALSE)
}
\arguments{
\item{gblup}{an object of the class gblup generated by \code{\link{gblup}}}

\item{ehat}{a logical value to indicate if residuals should be reported}

\item{sigma}{a logical value to indicate if variance component matrix should be included}

\item{fe}{a logical value to indicate if a fixed effects table should be included}
}
\value{
an object of the class summary.gblup with:
\itemize{
   \item {\code{trait}} {name of the trait}
   \item {\code{vformula and formula}} {equations for random and fixed effects}
   \item {\code{llik}} {log-likelihood at convergence}
   \item {\code{cycle}} {number of iterations}
   \item {\code{uhat}} {blup of individuals}
   \item{\code{sigma}} {estimated variance components and h2}
   \item{\code{coefm}} {estimated fixed effects, s.e. and test statistics}
}
}
\description{
A summary method for gblup objects
}
\seealso{
\code{\link{gblup}}
}

